+++ tests/api/v1/endpoints/test_connection_template_endpoints.py
@@ -4,7 +4,11 @@ from fideslib.models.client import ClientDetail
 from starlette.testclient import TestClient
 
 from fidesops.api.v1.scope_registry import CONNECTION_READ, CONNECTION_TYPE_READ
-from fidesops.api.v1.urn_registry import CONNECTION_TYPES, V1_URL_PREFIX
+from fidesops.api.v1.urn_registry import (
+    CONNECTION_TYPE_SECRETS,
+    CONNECTION_TYPES,
+    V1_URL_PREFIX,
+)
 
 
 class TestGetConnections:
@@ -57,3 +61,82 @@ class TestGetConnections:
         data = resp.json()
         assert data["total"] == 3
         assert data["items"] == ["outreach", "postgres", "redshift"]
+
+
+class TestGetConnectionSecretSchema:
+    @pytest.fixture(scope="function")
+    def base_url(self, oauth_client: ClientDetail, policy) -> str:
+        return V1_URL_PREFIX + CONNECTION_TYPE_SECRETS
+
+    def test_get_connection_secret_schema_not_authenticated(self, api_client, base_url):
+        resp = api_client.get(base_url.format(connection_type="sentry"), headers={})
+        assert resp.status_code == 401
+
+    def test_get_connection_secret_schema_forbidden(
+        self, api_client, base_url, generate_auth_header
+    ):
+        auth_header = generate_auth_header(scopes=[CONNECTION_READ])
+        resp = api_client.get(
+            base_url.format(connection_type="sentry"), headers=auth_header
+        )
+        assert resp.status_code == 403
+
+    def test_get_connection_secret_schema_not_found(
+        self, api_client: TestClient, generate_auth_header, base_url
+    ):
+        auth_header = generate_auth_header(scopes=[CONNECTION_TYPE_READ])
+        resp = api_client.get(
+            base_url.format(connection_type="connection_type_we_do_not_support"),
+            headers=auth_header,
+        )
+        assert resp.status_code == 404
+        assert (
+            resp.json()["detail"]
+            == "No connection type found with name 'connection_type_we_do_not_support'."
+        )
+
+    def test_get_connection_secret_schema_mongodb(
+        self, api_client: TestClient, generate_auth_header, base_url
+    ) -> None:
+        auth_header = generate_auth_header(scopes=[CONNECTION_TYPE_READ])
+        resp = api_client.get(
+            base_url.format(connection_type="mongodb"), headers=auth_header
+        )
+        assert resp.json() == {
+            "title": "MongoDBSchema",
+            "description": "Schema to validate the secrets needed to connect to a MongoDB Database",
+            "type": "object",
+            "properties": {
+                "url": {"title": "Url", "type": "string"},
+                "username": {"title": "Username", "type": "string"},
+                "password": {"title": "Password", "type": "string"},
+                "host": {"title": "Host", "type": "string"},
+                "port": {"title": "Port", "type": "integer"},
+                "defaultauthdb": {"title": "Defaultauthdb", "type": "string"},
+            },
+            "additionalProperties": False,
+        }
+
+    def test_get_connection_secret_schema_hubspot(
+        self, api_client: TestClient, generate_auth_header, base_url
+    ) -> None:
+        auth_header = generate_auth_header(scopes=[CONNECTION_TYPE_READ])
+        resp = api_client.get(
+            base_url.format(connection_type="hubspot"), headers=auth_header
+        )
+
+        assert resp.json() == {
+            "title": "hubspot_connector_example_schema",
+            "description": "Hubspot secrets schema",
+            "type": "object",
+            "properties": {
+                "hapikey": {"title": "Hapikey", "type": "string"},
+                "domain": {
+                    "title": "Domain",
+                    "default": "api.hubapi.com",
+                    "type": "string",
+                },
+            },
+            "required": ["hapikey"],
+            "additionalProperties": False,
+        }