+++ tests/fixtures/saas/hubspot_fixtures.py
@@ -16,9 +16,8 @@ from fidesops.models.datasetconfig import DatasetConfig
 from fidesops.schemas.saas.shared_schemas import HTTPMethod, SaaSRequestParams
 from fidesops.service.connectors import SaaSConnector
 from fidesops.util import cryptographic_util
-from fidesops.util.saas_util import format_body
+from fidesops.util.saas_util import format_body, load_config
 from tests.fixtures.application_fixtures import load_dataset
-from tests.fixtures.saas_example_fixtures import load_config
 from tests.test_helpers.saas_test_utils import poll_for_existence
 
 saas_config = load_toml(["saas_config.toml"])