+++ tests/fixtures/saas/salesforce_fixtures.py
@@ -15,8 +15,8 @@ from fidesops.models.connectionconfig import (
 )
 from fidesops.models.datasetconfig import DatasetConfig
 from fidesops.util import cryptographic_util
+from fidesops.util.saas_util import load_config
 from tests.fixtures.application_fixtures import load_dataset
-from tests.fixtures.saas_example_fixtures import load_config
 
 saas_config = load_toml(["saas_config.toml"])