+++ tests/fixtures/saas/zendesk_fixtures.py
@@ -14,9 +14,8 @@ from fidesops.models.connectionconfig import (
 )
 from fidesops.models.datasetconfig import DatasetConfig
 from fidesops.util import cryptographic_util
-from fidesops.util.saas_util import format_body
+from fidesops.util.saas_util import load_config
 from tests.fixtures.application_fixtures import load_dataset
-from tests.fixtures.saas_example_fixtures import load_config
 
 saas_config = load_toml(["saas_config.toml"])